Removed other docker files and change structure to java/com/asyncapi
