+++ outputClient/src/main/java/com/asyncapi/PostmanEchoWebSocketClient.java
@@ -8,6 +8,8 @@
 //////////////////////////////////////////////////////////////////////
 
 
+package com.asyncapi;
+
 import io.quarkus.websockets.next.WebSocketClient;
 import io.quarkus.websockets.next.WebSocketClientConnection;
 import io.quarkus.websockets.next.OnOpen;
