+++ packages/templates/clients/websocket/java/quarkus/components/ConnectorDependencies.js
@@ -5,6 +5,8 @@ export function ConnectorDependencies() {
     <Text>
       <Text>
         {`
+package com.asyncapi;
+
 import io.quarkus.websockets.next.WebSocketConnector;
 import io.quarkus.websockets.next.WebSocketClientConnection;
 import jakarta.inject.Inject;
