+++ packages/templates/clients/websocket/java/quarkus/template/README.md
@@ -1,4 +1,4 @@
-# websockets-next-quickstart
+# quarkus-websocket
 
 This project uses Quarkus, the Supersonic Subatomic Java Framework.
 
@@ -49,7 +49,7 @@ Or, if you don't have GraalVM installed, you can run the native executable build
 ./mvnw package -Dnative -Dquarkus.native.container-build=true
 ```
 
-You can then execute your native executable with: `./target/websockets-next-quickstart-1.0.0-SNAPSHOT-runner`
+ You can then execute your native executable with: `./target/quarkus-websocket-1.0.0-SNAPSHOT-runner`
 
 If you want to learn more about building native executables, please consult <https://quarkus.io/guides/maven-tooling>.
