+++ packages/templates/clients/websocket/java/quarkus/template/pom.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.acme</groupId>
-    <artifactId>websockets-next-quickstart</artifactId>
+    <groupId>com.asyncapi</groupId>
+    <artifactId>quarkus-websocket</artifactId>
     <version>1.0.0-SNAPSHOT</version>
 
     <properties>
