+++ packages/templates/clients/websocket/java/quarkus/template/src/main/java/com/asyncapi/config/config.js
